requirements.txt
streamlit
psycopg2-binary
pandas
sentence-transformers
torch
google-generativeai
